Don't run unicycler without Illumina reads.

+++ doit12.bash
@@ -6,28 +6,36 @@
 # Run unicycler (for a second opinion on the assembly)
 # ------------------------------------------------------------------------
 
-echo 1>&2 '# Running unicycler'
+if [ -z "${R1_FQ_GZ}" ] ; then
 
-rm -rf ${UNICYCLER}
-mkdir ${UNICYCLER}
+    echo 1>&2 '# No Illumina reads. Skipping unicycler.'
 
-unicycler -t ${THREADS} \
-	  -1 ${FASTP}/trimmed_R1.fastq.gz \
-	  -2 ${FASTP}/trimmed_R2.fastq.gz \
-	  -l ${FILTLONG}/filtered_nanopore.fastq.gz \
-	  -o ${UNICYCLER}
+else
+
+    echo 1>&2 '# Running unicycler'
+
+    rm -rf ${UNICYCLER}
+    mkdir ${UNICYCLER}
+
+    unicycler -t ${THREADS} \
+	      -1 ${FASTP}/trimmed_R1.fastq.gz \
+	      -2 ${FASTP}/trimmed_R2.fastq.gz \
+	      -l ${FILTLONG}/filtered_nanopore.fastq.gz \
+	      -o ${UNICYCLER}
+
+    echo 1>&2 '# Running DNADiff'
+
+    mkdir ${UNICYCLER}/dnadiff
+    cp ${DATA}/assembly.fasta ${UNICYCLER}/dnadiff/trycycler.fasta
+    cp ${UNICYCLER}/assembly.fasta ${UNICYCLER}/dnadiff/unicycler.fasta
+    cd ${UNICYCLER}/dnadiff
+    dnadiff trycycler.fasta unicycler.fasta
+    echo ''
+    cat out.report
+
+fi
 
 # ------------------------------------------------------------------------
-
-echo 1>&2 '# Running DNADiff'
-
-mkdir ${UNICYCLER}/dnadiff
-cp ${DATA}/assembly.fasta ${UNICYCLER}/dnadiff/trycycler.fasta
-cp ${UNICYCLER}/assembly.fasta ${UNICYCLER}/dnadiff/unicycler.fasta
-cd ${UNICYCLER}/dnadiff
-dnadiff trycycler.fasta unicycler.fasta
-echo ''
-cat out.report
 
 
 # ------------------------------------------------------------------------
